+++ src/constants/BUTTON.js
@@ -5,3 +5,9 @@ export const BUTTON_STATUS = Object.freeze({
   PRESSED: 'Pressed',
   FOCUS: 'Focus',
 });
+
+export const BUTTON_NAME = Object.freeze({
+  DELETE: '삭제하기',
+  EDIT: '편집하기',
+  COMPLETE_EDIT: '편집완료',
+});
