--- src/constants/PLACEHOLDER.js
@@ -1,4 +1,0 @@
-export const PLACEHOLDER = Object.freeze({
-  TO: '받는 사람 이름을 입력해 주세요.',
-  FROM: '이름을 입력해 주세요.',
-});
