+++ src/pages/PaperListPage.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import Carousel from '@components/list/Carousel';
 import styled from 'styled-components';
+
+import Carousel from '@components/list/Carousel';
 import Button from '@components/common/button/Button';
 import recipientsAPI from '@/api/recipientsAPI';
 import { API_RECIPIENTS } from '@constants/API';
